+++ components/Header.js
@@ -1,7 +1,105 @@
 import React, { useEffect, useRef, useState } from "react";
 import Image from "next/image";
-import styles from "../styles/LoginDropdown.module.css";
-import router from "next/router";
+import DropdownCard from "./LoginDropdown";
+
+// import styles from "../styles/LoginDropdown.module.css";
+// import router from "next/router";
+
+// const DropdownCard = ({ isLoggedIn, isUserLoggedIn, isMentorLoggedIn }) => {
+//   const [loginOption, setLoginOption] = useState(false);
+
+//   const handleLoginClick = () => {
+//     setLoginOption(!loginOption);
+//   };
+
+//   const mentorlogout = () => {
+//     localStorage.clear();
+//     setMentorLoggedIn(false);
+//     router.push("/");
+//     window.location.reload();
+//   };
+
+//   const userlogout = () => {
+//     localStorage.clear();
+//     setLoggedIn(false);
+//     router.push("/");
+//     window.location.reload();
+//   };
+
+//   return (
+//     <li>
+//       <div className={styles.loginOption}>
+//         {isLoggedIn || isUserLoggedIn || isMentorLoggedIn ? (
+//           <button onClick={handleLoginClick} className={styles.userName}>
+//             <img
+//               style={{
+//                 width: "35px",
+//                 height: "auto",
+//                 borderRadius: "50%",
+//               }}
+//               src={
+//                 localStorage.getItem("user_picture") ||
+//                 localStorage.getItem("mentor_picture") ||
+//                 "assets/img/icon/no-profile-picture.png"
+//               }
+//               alt="not found"
+//             />
+//           </button>
+//         ) : (
+//           <button
+//             className={styles.loginbutton}
+//             onClick={handleLoginClick}
+//           >
+//             Login
+//           </button>
+//         )}
+
+//         {loginOption && (
+//           <div className="login-optionslist">
+//             {isLoggedIn || isUserLoggedIn || isMentorLoggedIn ? (
+//               <>
+//                 <button
+//                   className="login-buttons"
+//                   style={{ marginTop: "20px" }}
+//                   onClick={() => {
+//                     if (isMentorLoggedIn) {
+//                       window.location.href = `/dashboard`;
+//                     } else {
+//                       window.location.href = `/`;
+//                     }
+//                   }}
+//                 >
+//                   Dashboard
+//                 </button>
+//                 <button
+//                   className="login-buttons"
+//                   onClick={() => {
+//                     if (isMentorLoggedIn) {
+//                       mentorlogout();
+//                     } else {
+//                       userlogout();
+//                     }
+//                   }}
+//                 >
+//                   Logout
+//                 </button>
+//               </>
+//             ) : (
+//               <>
+//                 <button className="login-buttons" onClick={handleLoginClick}>
+//                   <a href="/login">User</a>
+//                 </button>
+//                 <button className="login-buttons" onClick={handleLoginClick}>
+//                   <a href="/mentorLogin">Mentor</a>
+//                 </button>
+//               </>
+//             )}
+//           </div>
+//         )}
+//       </div>
+//     </li>
+//   );
+// };
 
 function Header({ isUserLoggedIn, navbarBackground }) {
   // localStorage.setItem('redirectUrl', window.location.href);
@@ -9,7 +107,6 @@ function Header({ isUserLoggedIn, navbarBackground }) {
   const [isMentorLoggedIn, setMentorLoggedIn] = useState(false);
   const [scrollY, setScrollY] = useState(0);
   const [navbarAppear, setNavbarAppear] = useState(false);
-  const [loginOption, setLoginOption] = useState(false);
 
   useEffect(() => {
     const userName = localStorage.getItem("user_name");
@@ -33,13 +130,7 @@ function Header({ isUserLoggedIn, navbarBackground }) {
     };
   }, []);
 
-  const handleLoginClick = () => {
-    setLoginOption(true);
-    setTimeout(() => {
-      setLoginOption(false);
-    }, 5000);
-  };
-
+  
   const menuToggle = () => {
     if (navbarAppear === true) {
       setNavbarAppear(false);
@@ -48,19 +139,6 @@ function Header({ isUserLoggedIn, navbarBackground }) {
     }
   };
 
-  const userlogout = () => {
-    localStorage.clear();
-    setLoggedIn(false);
-    router.push("/");
-    window.location.reload();
-  };
-
-  const mentorlogout = () => {
-    localStorage.clear();
-    setMentorLoggedIn(false);
-    router.push("/");
-    window.location.reload();
-  };
 
   return (
     <div className="header-area header-transparent">
@@ -108,93 +186,11 @@ function Header({ isUserLoggedIn, navbarBackground }) {
                         <li>
                           <a href="/contact">Contact</a>
                         </li>
-
-                        {isLoggedIn || isUserLoggedIn || isMentorLoggedIn ? (
-                          <li>
-                            <div className={styles.loginOption}>
-                              <button
-                                onClick={handleLoginClick}
-                                className={styles.userName}
-                              >
-                                <img
-                                  style={{
-                                    width: "35px",
-                                    height: "auto",
-                                    borderRadius: "50%",
-                                    backgroundColor: "ActiveBorder",
-                                  }}
-                                  src={
-                                    localStorage.getItem("user_picture") ||
-                                    localStorage.getItem("mentor_picture") ||
-                                    "assets/img/icon/no-profile-picture.png"
-                                  }
-                                  alt="not found"
-                                />
-                              </button>
-
-                              {loginOption && (
-                                <div className="login-optionslist">
-                                  <button
-                                    className="login-buttons"
-                                    // style={{ marginTop: "20px" }}
-                                    onClick={() => {
-                                      if (isMentorLoggedIn) {
-                                        window.location.href = `/dashboard`;
-                                      } else {
-                                        window.location.href = `/`;
-                                      }
-                                    }}
-                                  >
-                                    Dashboard
-                                  </button>
-                                  <button
-                                    className="login-buttons"
-                                    onClick={() => {
-                                      if (isMentorLoggedIn) {
-                                        mentorlogout();
-                                      } else {
-                                        userlogout();
-                                      }
-                                    }}
-                                  >
-                                    Logout
-                                  </button>
-                                </div>
-                              )}
-                            </div>
-                          </li>
-                        ) : (
-                          <li>
-                            <div className={styles.loginOption}>
-                              <button
-                                className={styles.loginbutton}
-                                onClick={handleLoginClick}
-                              >
-                                Login
-                              </button>
-                              {loginOption && (
-                                <div className="login-optionslist">
-                                  <button
-                                    className="login-buttons"
-                                    onClick={() => {
-                                      window.location.href = `/login`;
-                                    }}
-                                  >
-                                    User
-                                  </button>
-                                  <button
-                                    className="login-buttons"
-                                    onClick={() => {
-                                      window.location.href = `/mentorLogin`;
-                                    }}
-                                  >
-                                    Mentor
-                                  </button>
-                                </div>
-                              )}
-                            </div>
-                          </li>
-                        )}
+                        <DropdownCard
+                          isLoggedIn={isLoggedIn}
+                          isUserLoggedIn={isUserLoggedIn}
+                          isMentorLoggedIn={isMentorLoggedIn}
+                        />
                       </ul>
                     </nav>
                   </div>
